fix: export mapping nonevm with solana bridge txs to a hook
