+++ ui/components/app/transaction-list/transaction-list.component.js
@@ -20,20 +20,20 @@ import {
   nonceSortedPendingTransactionsSelector,
 } from '../../../selectors/transactions';
 import { getCurrentChainId } from '../../../../shared/modules/selectors/networks';
-import { NETWORK_TO_NAME_MAP } from '../../../../shared/constants/network';
-import { Numeric } from '../../../../shared/modules/Numeric';
 import {
   getSelectedAccount,
   ///: BEGIN:ONLY_INCLUDE_IF(build-main,build-beta,build-flask)
   getShouldHideZeroBalanceTokens,
   ///: END:ONLY_INCLUDE_IF
 } from '../../../selectors';
+///: BEGIN:ONLY_INCLUDE_IF(multichain)
+import useSolanaBridgeTransactionMapping from '../../../hooks/bridge/useSolanaBridgeTransactionMapping';
+///: END:ONLY_INCLUDE_IF
 import { useI18nContext } from '../../../hooks/useI18nContext';
 import TransactionListItem from '../transaction-list-item';
 import SmartTransactionListItem from '../transaction-list-item/smart-transaction-list-item.component';
 import { TOKEN_CATEGORY_HASH } from '../../../helpers/constants/transactions';
 import { SWAPS_CHAINID_CONTRACT_ADDRESS_MAP } from '../../../../shared/constants/swaps';
-import { selectBridgeHistoryForAccount } from '../../../ducks/bridge-status/selectors';
 import { isEqualCaseInsensitive } from '../../../../shared/modules/string-utils';
 import {
   Box,
@@ -204,143 +204,9 @@ export default function TransactionList({
     getSelectedAccountMultichainTransactions,
   );
 
-  // Log non-EVM transactions to see what's being shown for Solana accounts
-  console.log('Non-EVM transactions for selected account:', nonEvmTransactions);
-
-  // Get bridge transactions from the bridge status controller
-  const bridgeHistory = useSelector(selectBridgeHistoryForAccount);
-  console.log('Bridge history from status controller:', bridgeHistory);
-
-  // Log transaction signatures for direct comparison
-  console.log('==== SIGNATURES COMPARISON ====');
-
-  // Log non-EVM transaction signatures
-  console.log('Non-EVM transaction signatures:');
-  nonEvmTransactions?.transactions?.forEach((tx) => {
-    console.log(`NonEVM TX: ${tx.type}, Signature: ${tx.id}`);
-  });
-
-  // Log bridge transaction signatures
-  console.log('Bridge transaction signatures:');
-  Object.values(bridgeHistory || {}).forEach((bridgeTx) => {
-    console.log(
-      `Bridge TX: ${bridgeTx.txMetaId}, Signature: ${bridgeTx.status?.srcChain?.txHash}, Amount: ${bridgeTx.pricingData?.amountSent}`,
-    );
-  });
-
-  // Look for non-EVM transactions that match bridge transactions by signature/hash
-  const modifiedNonEvmTransactions = useMemo(() => {
-    if (!nonEvmTransactions?.transactions?.length || !bridgeHistory) {
-      return nonEvmTransactions;
-    }
-
-    // Create a map of bridge tx signatures to look up quickly
-    const bridgeTxSignatures = {};
-    Object.values(bridgeHistory).forEach((bridgeTx) => {
-      if (bridgeTx.status?.srcChain?.txHash) {
-        bridgeTxSignatures[bridgeTx.status.srcChain.txHash] = bridgeTx;
-      }
-    });
-
-    console.log('Bridge transaction signatures map:', bridgeTxSignatures);
-
-    // Create a modified copy of the transactions
-    const modifiedTransactions = nonEvmTransactions.transactions.map((tx) => {
-      // Check if this transaction's signature (id field) matches a bridge transaction
-      const txSignature = tx.id;
-
-      if (txSignature && bridgeTxSignatures[txSignature]) {
-        const matchingBridgeTx = bridgeTxSignatures[txSignature];
-        console.log('Found matching bridge transaction:', {
-          nonEvmTx: tx,
-          bridgeTx: matchingBridgeTx,
-        });
-
-        // Return a modified version of the transaction with bridge info
-        return {
-          ...tx,
-          // Change the type to bridge
-          type: 'bridge',
-          // Add bridge-specific fields
-          isBridgeTx: true,
-          // Include destination chain details
-          bridgeInfo: {
-            destChainId: matchingBridgeTx.quote?.destChainId,
-            destAsset: matchingBridgeTx.quote?.destAsset,
-            destTokenAmount: matchingBridgeTx.quote?.destTokenAmount,
-          },
-        };
-      }
-
-      // Return the original transaction unchanged
-      return tx;
-    });
-
-    return {
-      ...nonEvmTransactions,
-      transactions: modifiedTransactions,
-    };
-  }, [nonEvmTransactions, bridgeHistory]);
-
-  // Replace the original nonEvmTransactions with our modified version
-  // that correctly identifies bridge transactions
-  if (modifiedNonEvmTransactions !== nonEvmTransactions) {
-    console.log(
-      'Modified non-EVM transactions with bridge info:',
-      modifiedNonEvmTransactions,
-    );
-  }
-
-  // Check for exact matches between non-EVM transaction signatures and bridge transaction signatures
-  console.log('==== CHECKING FOR EXACT SIGNATURE MATCHES ====');
-
-  // Create a set of non-EVM transaction signatures for quick lookup
-  const nonEvmSignatures = new Set();
-  nonEvmTransactions?.transactions?.forEach((tx) => {
-    const signature = tx.id; // The signature is in the id field
-    if (signature) {
-      nonEvmSignatures.add(signature);
-    }
-  });
-
-  // Check bridge transactions for matching signatures
-  Object.values(bridgeHistory || {}).forEach((bridgeTx) => {
-    const bridgeSignature = bridgeTx.status?.srcChain?.txHash;
-    if (bridgeSignature && nonEvmSignatures.has(bridgeSignature)) {
-      console.log('EXACT SIGNATURE MATCH FOUND:', {
-        bridgeTxId: bridgeTx.txMetaId,
-        signature: bridgeSignature,
-        destChain: bridgeTx.quote?.destChainId,
-        destAsset: bridgeTx.quote?.destAsset?.symbol,
-      });
-    }
-  });
-
-  // Look for Solana bridge transactions in the bridge history
-  const solanaBridgeTxs = Object.values(bridgeHistory || {}).filter(
-    (historyItem) => {
-      // Check if any part of the transaction indicates it's a Solana transaction
-      const isSolanaBridge =
-        historyItem.quoteResponse?.isSolana === true ||
-        historyItem.quoteResponse?.srcChainId?.toString().includes('solana') ||
-        historyItem.quoteResponse?.quote?.srcChainId
-          ?.toString()
-          .includes('solana') ||
-        historyItem.quote?.srcChainId.toString().includes('1151111081099710'); // Solana chain ID
-
-      if (isSolanaBridge) {
-        console.log('Found Solana bridge transaction in history:', {
-          id: historyItem.txMetaId,
-          signature: historyItem.status?.srcChain?.txHash,
-          srcChainId: historyItem.quote?.srcChainId,
-        });
-      }
-
-      return isSolanaBridge;
-    },
-  );
-
-  console.log('Solana bridge transactions:', solanaBridgeTxs);
+  // Use our custom hook to map Solana bridge transactions with destination chain info
+  const modifiedNonEvmTransactions =
+    useSolanaBridgeTransactionMapping(nonEvmTransactions);
 
   const isSolanaAccount = useSelector(isSelectedInternalAccountSolana);
   ///: END:ONLY_INCLUDE_IF
@@ -590,38 +456,8 @@ export default function TransactionList({
                                 {`${t('to')} ${
                                   transaction.bridgeInfo.destAsset?.symbol
                                 } ${t('on')} ${
-                                  // First try to use MULTICHAIN_PROVIDER_CONFIGS for non-EVM chains
-                                  MULTICHAIN_PROVIDER_CONFIGS[
-                                    transaction.bridgeInfo.destChainId
-                                  ]?.nickname ||
-                                  // For EVM chains, convert decimal chain ID to hex and look up in NETWORK_TO_NAME_MAP
-                                  (() => {
-                                    try {
-                                      // Only attempt conversion if it's a number (EVM chain IDs are typically numbers)
-                                      if (
-                                        !isNaN(
-                                          Number(
-                                            transaction.bridgeInfo.destChainId,
-                                          ),
-                                        )
-                                      ) {
-                                        // Convert decimal to hex string with '0x' prefix
-                                        const hexChainId = new Numeric(
-                                          transaction.bridgeInfo.destChainId,
-                                          10,
-                                        ).toPrefixedHexString();
-
-                                        return (
-                                          NETWORK_TO_NAME_MAP[hexChainId] ||
-                                          null
-                                        );
-                                      }
-                                      return null;
-                                    } catch (e) {
-                                      return null;
-                                    }
-                                  })() ||
-                                  // Fall back to chain ID as last resort
+                                  // Use the pre-computed chain name from our hook, or fall back to chain ID
+                                  transaction.bridgeInfo.destChainName ||
                                   transaction.bridgeInfo.destChainId
                                 }`}
                               </Text>
