+++ ui/hooks/bridge/useSolanaBridgeTransactionMapping.ts
@@ -1,0 +1,95 @@
+import { useSelector } from 'react-redux';
+import { Numeric, NumericValue } from '../../../shared/modules/Numeric';
+import { NETWORK_TO_NAME_MAP } from '../../../shared/constants/network';
+import { MULTICHAIN_PROVIDER_CONFIGS } from '../../../shared/constants/multichain/networks';
+import { selectBridgeHistoryForAccount } from '../../ducks/bridge-status/selectors';
+import { Transaction } from '@metamask/keyring-api';
+
+/**
+ * Hook to map Solana bridge transactions with EVM destination info
+ * @returns Enhanced non-EVM transactions with bridging information
+ */
+export default function useSolanaBridgeTransactionMapping(
+  nonEvmTransactions:
+    | { transactions: Transaction[]; next: string | null; lastUpdated: number }
+    | undefined,
+) {
+  // Get bridge transactions from the bridge status controller
+  const bridgeHistory = useSelector(selectBridgeHistoryForAccount);
+
+  /**
+   * Gets a human-readable network name from a chain ID
+   * @param chainId - The chain ID to resolve (can be decimal or hex)
+   * @returns The network name or the original chain ID if not found
+   */
+  const getNetworkName = (chainId: NumericValue) => {
+    // First check if it's in the MULTICHAIN_PROVIDER_CONFIGS (for non-EVM chains)
+    let networkName = MULTICHAIN_PROVIDER_CONFIGS[chainId]?.nickname;
+
+    // If not found and it might be an EVM chain ID, convert to hex and check NETWORK_TO_NAME_MAP
+    if (!networkName && !isNaN(Number(chainId))) {
+      try {
+        // Convert decimal to hex with '0x' prefix
+        const hexChainId = new Numeric(chainId, 10).toPrefixedHexString();
+        networkName = NETWORK_TO_NAME_MAP[hexChainId];
+      } catch (e) {
+        // If conversion fails, just use the original chain ID
+        console.error('Error converting chain ID', e);
+      }
+    }
+
+    // Return the network name or the original chain ID if no mapping found
+    return networkName || chainId;
+  };
+
+  // Create a map of bridge transaction signatures for quick lookups
+  const bridgeTxSignatures = {};
+  if (bridgeHistory) {
+    Object.values(bridgeHistory).forEach((bridgeTx) => {
+      if (bridgeTx.status?.srcChain?.txHash) {
+        bridgeTxSignatures[bridgeTx.status.srcChain.txHash] = bridgeTx;
+      }
+    });
+  }
+
+  // No transactions to process
+  if (!nonEvmTransactions?.transactions?.length) {
+    return nonEvmTransactions;
+  }
+
+  // Create a modified copy with bridge info added
+  const modifiedTransactions = nonEvmTransactions.transactions.map((tx) => {
+    // The signature is in the id field for non-EVM transactions
+    const txSignature = tx.id;
+
+    // Check if this transaction signature matches a bridge transaction
+    if (txSignature && bridgeTxSignatures[txSignature]) {
+      const matchingBridgeTx = bridgeTxSignatures[txSignature];
+
+      // Return an enhanced version of the transaction with bridge info
+      return {
+        ...tx,
+        // Change the type to bridge
+        type: 'bridge',
+        // Add bridge-specific flags
+        isBridgeTx: true,
+        // Include destination chain details
+        bridgeInfo: {
+          destChainId: matchingBridgeTx.quote?.destChainId,
+          destChainName: getNetworkName(matchingBridgeTx.quote?.destChainId),
+          destAsset: matchingBridgeTx.quote?.destAsset,
+          destTokenAmount: matchingBridgeTx.quote?.destTokenAmount,
+        },
+      };
+    }
+
+    // Return the original transaction unchanged if not a bridge transaction
+    return tx;
+  });
+
+  // Return a modified copy of the original transactions with bridge info
+  return {
+    ...nonEvmTransactions,
+    transactions: modifiedTransactions,
+  };
+}
